Move `test_collidable_point_jacobians` to `test_api_contact`
